+++ src/Ch03/Braces.hs
@@ -14,11 +14,15 @@ module Ch03.Braces where
 
 -- p. 66
 
+-- added a signature to prevent [-Wmissing-signatures] warnings
+-- adding a signature removes [-Wtype-defaults] also
+bar :: Int
 bar = let a = 1
           b = 2
           c = 3
        in a + b + c
 
+foo :: Int
 foo = let {a = 1; b = 2;
         c = 3}
        in a + b + c