+++ ChangeLog.md
@@ -13,7 +13,8 @@ Fixed warnings
 * src/Ch03/BadPattern.hs        (added signature, and comment bad definition)
 * src/Ch03/BogusPattern.hs      (comment all definitions, defined a new function)
 * src/Ch03/BookStore.hs         (added signatures and comment first version accessor functions)
-* src/Ch03/Braces.hs            (added signatures)
+* src/Ch03/Braces.hs            (added signatures, removes -Wtypes-defaults also)
+* src/Ch03/Distinction.hs       (added signatures)
 * ChangeLog.md
 * package.yaml
 * realw.cabal